improve data receiver

+++ app/src/main/java/com/ucdavis/dart/cardiac_arrhythmia_app/Constants.java
@@ -8,5 +8,5 @@ import java.util.UUID;
 public class Constants {
     //"f741fa52-bd68-4f10-8529-934017a049b8"
     public static final UUID MY_UUID = UUID.fromString("00001101-0000-1000-8000-00805f9b34fb");
-    public static final Integer domain = 70;
+    public static final double domain = 70.0;
 }
